Locationをdefineに移動、MapObjectDataを空で作成
