+++ Classes/Datas/MapObject/MapObjectData.cpp
@@ -1,0 +1,9 @@
+//
+//  MapObjectData.cpp
+//  LastSupper
+//
+//  Created by 猪野凌也 on 2015/11/18.
+//
+//
+
+#include "Datas/MapObject/MapObjectData.h"
