+++ Classes/Datas/MapObject/MapObjectData.h
@@ -1,0 +1,17 @@
+//
+//  MapObjectData.h
+//  LastSupper
+//
+//  Created by 猪野凌也 on 2015/11/18.
+//
+//
+
+#ifndef MapObjectData_h
+#define MapObjectData_h
+
+#include "define.h"
+
+
+
+
+#endif /* MapObjectData_h */
